api-gateway/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
python-jose[cryptography]